% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagefinder.r
\name{fp}
\alias{fp}
\title{Searching for packages on CRAN}
\usage{
fp(...)
}
\arguments{
\item{...}{Arguments as in \code{\link{findPackage}()}.}
}
\description{
Shorter alias for function \code{\link{findPackage}()}.
}
\examples{
\donttest{
fp(c("meta", "regression"))
}
}
\author{
Joachim Zuckarelli \email{joachim@zuckarelli.de}
}
