% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagefinder.r
\name{go}
\alias{go}
\title{Showing information about a package}
\usage{
go(package, index = NULL, where.to = "manual")
}
\arguments{
\item{package}{Either the name of the package (capitalization does generally not matter) or the search result number shown in the results of \code{\link{findPackage}} (the number in the \code{GO} column).}

\item{index}{Either a path (or URL) to a search index, or a search index that is already loaded. If no index is provided, \code{findPackage} creates an ad hoc search index.}

\item{where.to}{Either \code{"manual"}, \code{"cran"} or \code{"install"}. With \code{where.to = "manual"} \code{go()} opens the package's PDF manual from the web, \code{where.to = "cran"} pulls up the package's CRAN website and \code{where.to = "install"} installs the package (including dependencies).}
}
\value{
No return value.
}
\description{
Allows to inspect a package found with \code{\link{findPackage}} by showing its manual (PDF) or go to its CRAN webpage.
}
\details{
\code{go()} is made to inspect a package found with \code{findPackage()} and to decide whether or not this package serves the intended purposes.
}
\author{
Joachim Zuckarelli \email{joachim@zuckarelli.de}
}
