% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagefinder.r
\name{packageDetails}
\alias{packageDetails}
\title{Showing information about a package}
\usage{
packageDetails(package, brief = FALSE, show.tip = TRUE, index = NULL)
}
\arguments{
\item{package}{Either the name of a package (capitalization does generally not matter) or the search result number shown in the results of \code{\link{findPackage}()} (the number in the \code{GO} column). Only one package is allowed here.}

\item{brief}{If \code{"TRUE"}, only short an long description as well as the maintainer of the package are shown, otherwise all available fields are displayed.}

\item{show.tip}{If \code{"TRUE"}, tips for getting additional information on the package are shown.}

\item{index}{Either a path (or URL) to a search index, or a search index that is already loaded. If no index is provided, \code{packageDetails()} creates an ad hoc search index.}
}
\value{
No return value.
}
\description{
Shows detailed CRAN information for a package.
}
\examples{
packageDetails("ggplot2")

}
\author{
Joachim Zuckarelli \email{joachim@zuckarelli.de}
}
