% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagefinder.r
\name{packagefinder}
\alias{packagefinder}
\title{Package 'packagefinder'}
\description{
Comfortable search for R packages on CRAN directly from the R console
}
\section{What is packagefinder? How does it benefit me?}{


Currently, there are more than 13,000 R package contributions on CRAN providing R with an unparalleled wealth of features. The downside of the large and increasing amount of packages is that it becomes increasingly difficult to find the right tools to tackle a specific problem. Unfortunately, CRAN does not provide any good search functionality.

\strong{packagefinder} is designed to search for CRAN packages right from the R console. The philosophy behind this package is that R users like using the R console and need a tool to do their day-to-day-work on CRAN without leaving their normal workspace, the console. In fact, the idea is that with \strong{packagefinder} you do not \emph{need} to leave the R console to work with CRAN effectively.

\strong{packagefinder} is developed to save you time and energy finding the right packages to work with, thereby making your work with R more productive.
}

\section{Where can I get more information?}{


\itemize{
\item Contact the \strong{author}:
\itemize{
\item E-mail: Joachim Zuckarelli (\email{joachim@zuckarelli.de})
\item Twitter: \href{https://twitter.com/jsugarelli}{@jsugarelli}
}
\item packagefinder \strong{introduction}: \href{http://www.zuckarelli.de/packagefinder/tutorial.html}{A Quick Tutorial}
\item packagefinder on \strong{GitHub}: \href{https://github.com/jsugarelli/packagefinder}{https://github.com/jsugarelli/packagefinder}
\item packagefinder video tutorial on \strong{YouTube}: \href{https://www.youtube.com/watch?v=66Mes6_hYno}{https://www.youtube.com/watch?v=66Mes6_hYno}
}
}

