% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagefinder.r
\name{whatsNew}
\alias{whatsNew}
\title{Staying up-to-date on CRAN packages}
\usage{
whatsNew(last.days = 0, brief = TRUE, index = NULL)
}
\arguments{
\item{last.days}{The length of the period (in days) for which package additions to CRAN shall be presented. \code{last.days=0} means only today's additions are shown.}

\item{brief}{Determines if all avalilable package description fields shall be shown (\code{brief=FALSE}) or only a summary covering the most important fields (\code{brief=TRUE}, the default).}

\item{index}{Either a path (or URL) to a search index, or a search index that is already loaded. If no index is provided, \code{whatsNew()} creates an ad hoc search index.}
}
\value{
Number of packages covered by the period specified in \code{last.days}.
}
\description{
Shows information on the latest package additions to CRAN.
}
\examples{
whatsNew(last.days = 3)

}
\author{
Joachim Zuckarelli \email{joachim@zuckarelli.de}
}
